refactor: rename to Main
