+++ wondrous-app/components/Common/Sidebar/index.tsx
@@ -1,0 +1,2 @@
+export { default as SideBarComponent } from './Main';
+export { default as DaoSidebar } from './Organization';
